requirements.txt
tqdm
numpy
beartype
jaxtyping
omegaconf
hydra-core
Pillow
imageio[ffmpeg]
opencv-python
torchmetrics[image]
lpips
tensorboard
ipykernel